backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
defusedxml==0.7.1
e2b==1.7.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.6.3
packaging==25.0
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
